+++ app/src/main/java/cos/mos/utils/ui/sideslip/SideLipFragment.java
@@ -50,7 +50,7 @@ public class SideLipFragment extends KFragment {
         rv.setLayoutManager(new LinearLayoutManager(context));
         adapter.setEmptyView(R.layout.lay_empty, rv);
         rv.addItemDecoration(new MyDividerDecoration(ContextCompat.getColor(context, R.color.fun_bg_white),
-            UScreen.dp2px(10), UScreen.dp2px(10), UScreen.dp2px(10)));
+            (int) UScreen.dp2px(10), (int) UScreen.dp2px(10), (int) UScreen.dp2px(10)));
         rv.setAdapter(adapter);
     }
